requirements.txt
deepspeed==0.11.1
numpy==1.26.1
PyYAML==6.0.1
torch==2.0.0
tqdm==4.66.1
transformers==4.33.1